backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
asyncpg
python-jose==3.3.0
cryptography>=41.0.0
python-multipart==0.0.6
pydantic>=1.10.0,<2.0.0
python-dotenv==1.0.0